SEMACHv3 batch evaluation

Wrapper for streamlined evaluation of larger sets of respiration measurements. Requires proper folder structure for Chambers / Measurement data.

Usage

SEMACHv3_batch_evaluation(
  dirpath =
    "//zfs1.hrz.tu-freiberg.de/fak3ibf/Hydropedo/field_data/data/SEMACHv3/SEMACHv3.1.0/",
  metadata_path = paste0(dirpath, "metadata.csv"),
  ch_list = paste0("Ch-", c(1:9)),
  cutoff_start = 181,
  cutoff_end = 99999,
  from = "2024-01-01",
  to = "3000-01-01"
)

Arguments

dirpath

Path to parent folder. ./dirpath should contain folders for each chamber named as stated in ch_list. See details.

metadata_path

Path to a .csv file containing slope and offset for each chamber system. Must contain at least 4 columns: System - the chamber system id used in the folder structure, SCD30_slope - the slope correction for SCD30 data, SCD30_offset - the offset correction for SCD30 data, height_offset - height difference of system volume. Note that this can here only be changed for each subfolder, not each ring / measurement site. SCD30 slope and offset can be calculated by SEMACHv3_SCD30calib. If data should be processed without corrections, entries of SCD30_slope have to be set to 1, and both SCD30_offset and height_offset to 0.

ch_list

Character vector of individual chamber systems. Must be same as folder names in dirpath and entries column System in cal_lookup_path.

cutoff_start

The first n seconds of the measurement will be disregarded (equilibration period, 60s recommended).

cutoff_end

Measurement will only be used to this point (in s). Long measurements might have non-linear behaviour.

from

Only evaluate chamber data recored after this date.

to

Only evaluate chamber data recored before this date.

Details

Currently, no way of attributing an site or measurement id to chamber measurements is implemented. The only way to identify measurements is the timestamp that is used as filename. Therefore, a clear folder structure is paramount to differentiate between chamber systems and allow the assignment of measurement to individual sites. Data therefore recommended to be stored in a folder structure as follows:

 ./dirpath

   /Ch-1

       /date1

           measurement1.csv

           measurement2.csv

           ...

   /Ch-2

   ...

   /metadata.csv