load Pri8800 data

load different files created by Pri8800 (both inidvidual slopes and final output files).

Usage

load_pri8800_data(
  folder,
  contains_folders = T,
  calculate_flux_manual = F,
  height_list = rep(NA, 25),
  wt_list = rep(NA, 25),
  V_head = 6.5 * (3.85/2)^2 * pi,
  V_sys = 190,
  V_delta = 0,
  start_cutoff = 60,
  end_cutoff = 450
)

Arguments

folder

Path to folder containing measurement files.

contains_folders

Boolean. Recursive read in or not.

calculate_flux_manual

Boolean. Should flux be re-calculated from slope data? This might be slow.

wt_list

List of soil sample weights in each vessel in grams. Vector of length 25 required, corresponding to sample positions.

V_head

Air volume of the plastic cylinder heads in mL. Default is 6.5 cm x (3.85 cm)^2 x pi = 75.67 mL

V_sys

System volume (Tubing etc). Default is 182 mL + 8 mL = 190 mL.

V_delta

Additional volume tuning argument in mL. Default = 0.

start_cutoff

Data is used for evaluation from this time (sec) after after start of recording. Default 60s

end_cutoff

Data is used for evaluation to this time (sec) after after start of recording. Default 450s

heigth_list

Sample height in the cylinder in cm

Note

Known issue: First sample of a run is not evaluated.