Flux calculator (legacy version)

Legacy flux calculation for PriEco8800 incubator. This function might be unstable. Use at own risk.

Usage

calc_flux_manual(
  dataset,
  V_ch_list = c(NA),
  A_ch = NA,
  wt_list = c(NA),
  start_cutoff = 60,
  end_cutoff = 360
)

Arguments

dataset

Flux measurement data for slope calculation as loaded by load_pri8800_data.

V_ch_list

Total system volume (accumulation volume) of each incubation vessel. If NA defaults to 1. cubic meter

A_ch

Default base area - For compatibility with field chambers. If NA defaults to 1. square meter

wt_list

List of soil sample weights in each vessel. grams

start_cutoff

Data is used for evaluation from this time (sec) after after start of recording. Default 60s

end_cutoff

Data is used for evaluation to this time (sec) after after start of recording. Default 360s

Note

optimise potentially: Function for individual meas, in main function, group by run,round,sample; summarise